+++ src/app/seasons/[slug]/users/[userId]/page.tsx
@@ -1,5 +1,4 @@
 import { Card } from "@/components/ui/card";
-import { SocialBadge } from "@/components/ui/social-badge";
 import { UserMissionAchievements } from "@/features/user-achievements/components/user-mission-achievements";
 import { getUserRepeatableMissionAchievements } from "@/features/user-achievements/services/achievements";
 /**
@@ -24,6 +23,7 @@ import {
 } from "@/features/user-activity/services/timeline";
 import { UserBadges } from "@/features/user-badges/components/user-badges";
 import Levels from "@/features/user-level/components/levels";
+import { SocialBadge } from "@/features/user-profile/components/social-badge";
 import { UserSeasonHeader } from "@/features/user-season/components/user-season-header";
 import { UserSeasonHistory } from "@/features/user-season/components/user-season-history";
 import { getSeasonBySlug, getUserSeasonHistory } from "@/lib/services/seasons";
