+++ src/app/users/[id]/page.tsx
@@ -1,5 +1,4 @@
 import { Card } from "@/components/ui/card";
-import { SocialBadge } from "@/components/ui/social-badge";
 import { UserMissionAchievements } from "@/features/user-achievements/components/user-mission-achievements";
 import { getUserRepeatableMissionAchievements } from "@/features/user-achievements/services/achievements";
 import UserDetailActivities from "@/features/user-activity/components/user-detail-activities";
@@ -22,6 +21,7 @@ import { UserBadges } from "@/features/user-badges/components/user-badges";
  * - 初期データをクライアントコンポーネントに渡してSSR最適化
  */
 import Levels from "@/features/user-level/components/levels";
+import { SocialBadge } from "@/features/user-profile/components/social-badge";
 import { UserSeasonHistory } from "@/features/user-season/components/user-season-history";
 import {
   getCurrentSeasonId,
