+++ src/components/collapsible-info.tsx
