+++ src/features/missions/components/difficulty-badge.tsx
