+++ src/features/missions/components/mission-icon.tsx
