+++ src/features/onboarding/components/onboarding-mission-details.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { DifficultyBadge } from "@/components/ui/difficulty-badge";
-import { MissionIcon } from "@/components/ui/mission-icon";
+import { DifficultyBadge } from "@/features/missions/components/difficulty-badge";
+import { MissionIcon } from "@/features/missions/components/mission-icon";
 import type { MockMission } from "@/features/onboarding/types/types";
 import { BUTTON_TEXT, STYLE_CLASSES } from "../constants/constants";
 import { sanitizeHtml } from "../utils/utils";
