+++ src/features/user-level/components/level-progress.tsx
@@ -1,4 +1,4 @@
-import { ProgressBarSimple } from "@/components/ui/progress-bar-simple";
+import { ProgressBarSimple } from "@/features/user-level/components/progress-bar-simple";
 import type { UserLevel } from "@/features/user-level/types/level-types";
 import React from "react";
