+++ src/features/user-level/components/progress-bar-animated.tsx
