+++ src/features/user-level/components/progress-bar-simple.tsx
