+++ src/features/user-level/components/progress-circle.tsx
