+++ src/features/user-profile/components/social-badge.tsx
@@ -1,4 +1,4 @@
-import { Badge } from "./badge";
+import { Badge } from "@/components/ui/badge";
 
 interface SocialBadgeProps {
   username: string;
